Auth middleware creation, and other changes.

+++ src/app/api/admin/delete-product/route.js
@@ -1,4 +1,5 @@
 import connectToDB from "@/database";
+import AuthUser from "@/middleware/AuthUser";
 import Product from "@/models/products";
 import { NextResponse } from "next/server";
 
@@ -7,23 +8,32 @@ export const dyanmic = "force-dynamic";
 export async function DELETE(req) {
   try {
     await connectToDB();
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get("id");
+    const isAuthUser = await AuthUser(req);
+    
+    if (isAuthUser?.role === "admin") {
+      const { searchParams } = new URL(req.url);
+      const id = searchParams.get("id");
 
-    if (!id)
-      return NextResponse.json({ success: false, message: "Id missing." });
+      if (!id)
+        return NextResponse.json({ success: false, message: "Id missing." });
 
-    const deletedProduct = await Product.findOneAndDelete(id);
+      const deletedProduct = await Product.findOneAndDelete(id);
 
-    if (deletedProduct) {
-      return NextResponse.json({
-        success: true,
-        message: "Product Deleted successfully",
-      });
+      if (deletedProduct) {
+        return NextResponse.json({
+          success: true,
+          message: "Product Deleted successfully",
+        });
+      } else {
+        return NextResponse.json({
+          success: false,
+          message: "Failed to delete product.",
+        });
+      }
     } else {
       return NextResponse.json({
         success: false,
-        message: "Failed to delete product.",
+        message: "You are not autorized !",
       });
     }
   } catch (error) {
